Fix bug where multiple Pause Overlays can be created
The problem still exists, I do not know how to fix the problem where we start loading the Pause scene but it has not finished loading yet. I suspect that this cannot be a user triggered thing, or I have a debouncer / wrapper object. I D K its a small enough edge case that I will figure this out later